Add a first prototype
With this, it's possible to set up a train track in the Godot editor,
using a TileMapLayer. On _ready, it will place areas on top of certain
tiles that issue a direction change for any object entering them.

At runtime, the player can indirectly control a train by clicking on a
switch in the track to toggle its direction (displayed in green).