Remodel 'player' scene into abstract 'train' scene
Considering the player has no direct control over the trains, this makes
sense. With that there can be more than one train in the game now.

Also, adds a train spawner scene that adds a new train to the game every
5 seconds (the interval is configurable via an export variable).

Trains are now of type CharacterBody2D and can collide with each other,
which (for now) will permanently remove both colliding trains.

The train speed was halved.